% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTransmission.R
\name{cases.plot}
\alias{cases.plot}
\title{Generate summary table and barplots (optional) of asymptomatic & symptomatic cases at given times since pandemic starts}
\usage{
cases.plot(transmatrix, times, plot = TRUE)
}
\arguments{
\item{transmatrix}{information matrix of transmission including infection time, recovery/dead time, number of people infected by current individual, current node number of current individual, node number of individual who infected the current person (ancestor), asymptomatic information, diagnosed time and death information of current person}

\item{times}{a vector of truncated times (in days) during pandemic}

\item{plot}{a logical parameter indicating whether to produce barplot or not (It's recommended to set to FALSE if vector "times" has a length longer than 15)}
}
\value{
barplots (optional) and summary tables illustrating number of cases of six groups (asymptomatic_infectious, asymptomatic_recovered, symptomatic_latent, symptomatic_diagnosed, symptomatic_recovered, dead) by truncated times.
}
\description{
Generate summary table and barplots (optional) of asymptomatic & symptomatic cases at given times since pandemic starts
}
