% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTransmission.R
\name{distNodeToroot}
\alias{distNodeToroot}
\title{Calculate the distance/time of input node from the first infected individual}
\usage{
distNodeToroot(node, nodematrix)
}
\arguments{
\item{node}{a number of node}

\item{nodematrix}{a matrix that describes the relationships among nodes and corresponding branch lengths}
}
\value{
distance/time of input node from the first infected individual
}
\description{
Calculate the distance/time of input node from the first infected individual
}
