% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTransmission.R
\name{findInfections}
\alias{findInfections}
\title{Find the infected individual and its next generation}
\usage{
findInfections(node, transmatrix)
}
\arguments{
\item{node}{node number of infected individual}

\item{transmatrix}{information matrix of transmission}
}
\value{
a transmission matrix including infection time, removal time and node number of current individual and the individuals infected by the input node.
}
\description{
Takes in transmission matrix, node of infected individual and return the information matrix of input node of infected individual as well as the individuals directly infected by it.
}
