% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTransmission.R
\name{modelTreefromTransmission}
\alias{modelTreefromTransmission}
\title{Generate transmission tree in phylip format}
\usage{
modelTreefromTransmission(transmatrix)
}
\arguments{
\item{transmatrix}{information matrix of transmission}
}
\value{
transmission tree in phylip format
}
\description{
Generate transmission tree in phylip format
}
