% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTransmission.R
\name{plotTransNet}
\alias{plotTransNet}
\title{Plot transmission network}
\usage{
plotTransNet(
  transmatrix,
  index_sym_latent = integer(0),
  index_sym_diagnosed,
  index_sym_recovered,
  index_asy_infectious = integer(0),
  index_asy_recovered,
  index_dead
)
}
\arguments{
\item{transmatrix}{information matrix of transmission including infection time, recovery/dead time, number of people infected by current individual, current node number of current individual, node number of individual who infected the current person (ancestor), asymptomatic information, diagnosed time and death information of current person}

\item{index_sym_latent}{index of symptomatic individuals who are in latent period}

\item{index_sym_diagnosed}{index of symptomatic individuals who have been diagnosed}

\item{index_sym_recovered}{index of symptomatic individuals who have recovered}

\item{index_asy_infectious}{index of asymptomatic individuals who are still infectious (not recovered).}

\item{index_asy_recovered}{index of asymptomatic individuals who have recovered}

\item{index_dead}{index of individuals who was dead from the pandemic}
}
\value{
a plot of transmission network
}
\description{
Takes in transmission matrix and plot a transmission network
}
