% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTransmission.R
\name{plotTree}
\alias{plotTree}
\title{Plot phylogenetic/transmission tree}
\usage{
plotTree(tree)
}
\arguments{
\item{tree}{tree in phylip format}
}
\value{
phylogenetic tree plot
}
\description{
Plot phylogenetic/transmission tree
}
