% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTransmission.R
\name{Runif}
\alias{Runif}
\title{Generate random numbers from uniform distribution with input parameters in vector format}
\usage{
Runif(input)
}
\arguments{
\item{input}{a vector containing 3 elements: amount of random numbers to be generated, lower bound and upper bound of uniform distribution}
}
\value{
a vector of generated random numbers
}
\description{
Generate random numbers from uniform distribution with input parameters in vector format
}
