% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTransmission.R
\name{simCoalModeltree}
\alias{simCoalModeltree}
\title{Simulate phylogenetic/coalescent tree from transmission tree}
\usage{
simCoalModeltree(ntree, modeltree, theta)
}
\arguments{
\item{ntree}{number of coalescent tree that one wants to generate}

\item{modeltree}{transmission tree in phylip format}

\item{theta}{product of effective sample size and mutation rate per day}
}
\value{
coalescent tree in phylip format
}
\description{
Simulate phylogenetic/coalescent tree from transmission tree
}
