% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTransmission.R
\name{simTransmission}
\alias{simTransmission}
\title{Simulate a transmission tree considering asymptomatic/unsampled individulas}
\usage{
simTransmission(
  inf_rate,
  inf_rate_time = 0,
  diag_rate,
  rec_rate_asy,
  rec_rate_sym,
  death_prop,
  death_rate,
  inf_rate_diag,
  asy_prop,
  ngeneration
)
}
\arguments{
\item{inf_rate}{per day infective rate per infected capita; could be a constant or vector}

\item{inf_rate_time}{time points corresponds to the start time of each infectious rate when it is not a constant}

\item{diag_rate}{per day diagnosis rate per capita for symptomatic/sampled individuals; constant}

\item{rec_rate_asy}{per day recovery rate per capita for asymptomatic/unsampled individulas; constant}

\item{rec_rate_sym}{per day recovery rate per capita for symptomatic/unsampled individulas after diagnosis; constant}

\item{death_prop}{proportion of dead cases among diagnosed people; constant}

\item{death_rate}{death rate per day per capita (inverse of the average time/days to death); constant}

\item{inf_rate_diag}{infection rate per day per capita for symptomatic individuals after diagnosis; constant}

\item{asy_prop}{aysmptomatic propotion among infected individuals; constant}

\item{ngeneration}{number of generation that get infected counted from the first infected individual}
}
\value{
transmission matrix including infection time, recovery/dead time, number of people infected by current individual, current node number of current individual, node number of individual who infected the current person (ancestor), asymptomatic information, diagnosed time and death information of current person
}
\description{
Takes in infection rate, recovery rate of symptomatic and asymptomatic group, asymptomatic proportion among infeted people and number of generation to generate a transmission tree
}
