% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTransmission.R
\name{subTree}
\alias{subTree}
\title{Simulate sub-transmission tree from one ancestor to its next generation}
\usage{
subTree(inf_time, rec_time, nodes)
}
\arguments{
\item{inf_time}{a vector of infection times}

\item{rec_time}{a vector of recovery times}

\item{nodes}{nodes number of infected individuals}
}
\value{
sub-transmission tree in phylip format
}
\description{
Simulate sub-transmission tree from one ancestor to its next generation
}
