% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTransmission.R
\name{timeTruncate}
\alias{timeTruncate}
\title{Truncate transmission matrix by a given time and produce total number and index of cases belong to each category.}
\usage{
timeTruncate(transmatrix, time)
}
\arguments{
\item{transmatrix}{information matrix of transmission including infection time, recovery/dead time, number of people infected by current individual, current node number of current individual, node number of individual who infected the current person (ancestor), asymptomatic information, diagnosed time and death information of current person}

\item{time}{a truncated time}
}
\value{
a time truncated transmission matrix along with total number and index of Symptomatic_latent, Symptomatic_diagnosed, Symptomatic_recovered, Asymptomatic_infectious, Asymptomatic_recovered and dead cases at the truncated time
}
\description{
Takes in transmission matrix and trucated it at given time. Then generate the truncated transmission matrix and latent/infectious and recovered cases for both syptomatic and asymptomatic groups, as well as diagnosed and dead cases by a truncated/given time
}
