% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTransmission.R
\name{toAdj}
\alias{toAdj}
\title{Generate transmission matrix that can be used to plot transmission network}
\usage{
toAdj(transmatrix)
}
\arguments{
\item{transmatrix}{information matrix of transmission including infection time, recovery/dead time, number of people infected by current individual, current node number of current individual, node number of individual who infected the current person (ancestor), asymptomatic information, diagnosed time and death information of current person}
}
\value{
network structure matrix
}
\description{
Takes in transmission matrix and convert it to a matrix giving network structure
}
