% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTransmission.R
\name{transCurrent}
\alias{transCurrent}
\title{Generate the first infected individual in a transmission tree with consideration of asyptomatic/unsampled individuals}
\usage{
transCurrent(
  inftime = 0,
  inf_rate,
  inf_rate_timeInt,
  diag_rate,
  rec_rate_asy,
  rec_rate_sym,
  death_prop,
  death_rate,
  inf_rate_diag,
  asy_prop,
  currentnode,
  ancestor
)
}
\arguments{
\item{inftime}{the start time of infection/outbreak}

\item{inf_rate}{per day infective rate per infected capita; could be a constant or vector}

\item{inf_rate_timeInt}{time intervals correspond to the start and end time of each infectious rate; matrix}

\item{diag_rate}{per day diagnosis rate per capita for symptomatic/sampled individuals; constant}

\item{rec_rate_asy}{per day recovery rate per capita for asymptomatic/unsampled individulas; constant}

\item{rec_rate_sym}{per day recovery rate per capita for symptomatic/unsampled individulas after diagnosis; constant}

\item{death_prop}{proportion of dead cases among diagnosed people; constant}

\item{death_rate}{death rate per day per capita (inverse of the average time/days to death); constant}

\item{inf_rate_diag}{infection rate per day per capita for symptomatic individuals after diagnosis; constant}

\item{asy_prop}{aysmptomatic propotion among infected individuals; constant}

\item{currentnode}{current node number of infected individual}

\item{ancestor}{node number of individual who infected the current person}
}
\value{
a list containing transmission information vector and input parameter vector to generate random numbers/times of next generation from uniform distribution
}
\description{
Generate the first infected individual in a transmission tree with consideration of asyptomatic/unsampled individuals
}
