% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTransmission.R
\name{transNext}
\alias{transNext}
\title{Generate transmission information of next generation with consideration of asyptomatic/unsampled individuals}
\usage{
transNext(
  currentmatrix,
  inf_rate,
  inf_rate_timeInt,
  diag_rate,
  rec_rate_asy,
  rec_rate_sym,
  death_prop,
  death_rate,
  inf_rate_diag,
  asy_prop,
  unif_input,
  ntrans
)
}
\arguments{
\item{currentmatrix}{information matrix of current generation}

\item{inf_rate}{per day infective rate per infected capita; could be a constant or vector}

\item{inf_rate_timeInt}{time interval correspond to the start and end time of each infectious rate; matrix}

\item{diag_rate}{per day diagnosis rate per capita for symptomatic/sampled individuals; constant}

\item{rec_rate_asy}{per day recovery rate per capita for asymptomatic/unsampled individulas; constant}

\item{rec_rate_sym}{per day recovery rate per capita for symptomatic/unsampled individulas after diagnosis; constant}

\item{death_prop}{proportion of dead cases among diagnosed people; constant}

\item{death_rate}{death rate per day per capita (inverse of the average time/days to death); constant}

\item{inf_rate_diag}{infection rate per day per capita for symptomatic individuals after diagnosis; constant}

\item{asy_prop}{aysmptomatic propotion among infected individuals; constant}

\item{unif_input}{a n*3 matrix containing input parameters to generate random numbers/times from uniform distribution}

\item{ntrans}{number of infected individuals from the begining of outbreak (first generation) to the current generation}
}
\value{
a list containing transmission information matrix and input parameter matrix to generate random numbers/times of next generation from uniform distribution
}
\description{
Generate transmission information of next generation with consideration of asyptomatic/unsampled individuals
}
